started to remove references to separate Vehicle components, and integrate them with Asset
