+++ ams-ng/src/app/core/vehicle/vehicle-add/vehicle-add.component.html
@@ -1,4 +1,4 @@
-<app-menu></app-menu>
+<!-- <app-menu></app-menu> -->
 <div class="MainContent form-horizontal">
 
 	<div><span class="PageTitle">Vehicle Add</span></div>
@@ -6,6 +6,46 @@
 	<div style="text-align: center;">
 
 	<form *ngIf="vehicle" style="display: inline-block; width: 70%; padding: 30px 30px 0 0; border: 1px solid silver;">
+
+	<div class='form-group'>
+    	<label class='col-md-3 control-label' for="asset-vendor">Vendor</label>
+    	<div class='col-md-9'>
+   			<input class='form-control' name='Vendor' type= 'text' id='asset-vendor' 
+   			[(ngModel)]='asset.Vendor' placeholder="Vendor">
+   		</div>
+   </div>
+
+	<div class='form-group'>
+		<label class='col-md-3 control-label' for="asset-name">Name</label>
+		<div class='col-md-9'>
+			<input class='form-control' name='Name' type= 'text' id='asset-name' 
+			[(ngModel)]='asset.Name' placeholder="Name">
+		</div>
+	</div>
+
+	<div class='form-group'>
+		<label class='col-md-3 control-label' for="asset-price">Price</label>
+		<div class='col-md-9'>
+			<input class='form-control' name='Price' type= 'text' id='asset-price' 
+			[(ngModel)]='asset.Price' placeholder="Price">
+		</div>
+	</div>
+
+	<div class='form-group'>
+		<label class='col-md-3 control-label' for="asset-unit">Unit</label>
+		<div class='col-md-9'>
+			<input class='form-control' name='Unit' type= 'text' id='asset-unit' 
+			[(ngModel)]='asset.Unit' placeholder="Unit">
+		</div>
+	</div>
+
+	<div class='form-group'>
+		<label class='col-md-3 control-label' for="asset-photopath">PhotoPath</label>
+		<div class='col-md-9'>
+			<input class='form-control' name='PhotoPath' type= 'text' id='asset-photopath' 
+			[(ngModel)]='asset.PhotoPath' placeholder="PhotoPath">
+		</div>
+	</div>
 
 	<div class='form-group' >
 		<label class='col-md-3  control-label' for="vehicle-make">Make</label>
