+++ ams-ng/src/app/core/vehicle/vehicle-detail/vehicle-detail.component.html
@@ -1,4 +1,5 @@
-<app-menu></app-menu>
+<!-- <app-menu></app-menu> -->
+<!-- We want to combine the asset information and vehicle information here, not on the list -->
 <div class="MainContent form-horizontal">
 
 	<div>
