+++ ams-ng/src/app/core/vehicle/vehicle-edit/vehicle-edit.component.html
@@ -1,4 +1,4 @@
-<app-menu></app-menu>
+<!-- <app-menu></app-menu> -->
 <div class="MainContent form-horizontal">
 
 	<div><span class="PageTitle">Vehicle Edit</span></div>
