+++ ams-ng/src/app/core/vehicle/vehicle-list/vehicle-list.component.html
@@ -1,6 +1,6 @@
-<app-menu></app-menu>
+<!-- <app-menu></app-menu> -->
 <div class="MainContent form-horizontal">
-
+<!-- We want to combine the asset information and vehicle information on the detail, not here -->
 	<div>
 		<span class="PageTitle">Vehicles</span>
 		<span class="CreateNew"><a routerLink="/vehicles/add">Add New</a></span>
