+++ ams-ng/src/app/core/vehicle/vehicle-list/vehicle-list.component.ts
@@ -2,8 +2,10 @@ import { Component, OnInit } from '@angular/core';
 
 import 'rxjs/add/operator/toPromise';
 
-import { Vehicle } from '../../models/Vehicle';
+import { Vehicle } from '../../../shared/models/Vehicle';
 import { VehicleService } from '../../../shared/services/vehicle.service';
+import { VehicleAsset } from '../../../shared/models/VehicleAsset';
+import { Asset } from '../../../shared/models/asset';
 
 @Component({
   selector: 'app-vehicle-list',
